Remove unwanted code and relocate the style files
